+++ rasa_nlu/evaluate.py
@@ -81,20 +81,28 @@ def plot_confusion_matrix(cm, classes,
     plt.xlabel('Predicted label')
 
 
-def get_evaluation_table(test_y, preds, log=True):  # pragma: no cover
+def log_evaluation_table(test_y, preds):  # pragma: no cover
+    """Logs the sklearn evaluation metrics"""
+    report, precision, f1, accuracy = get_evaluation_table(test_y, preds)
+
+    logger.info("Intent Evaluation Results")
+    logger.info("F1-Score:  {}".format(f1))
+    logger.info("Precision: {}".format(precision))
+    logger.info("Accuracy:  {}".format(accuracy))
+    logger.info("Classification report: \n{}".format(report))
+
+
+def get_evaluation_table(test_y, preds):  # pragma: no cover
+    """Computes the f1, precision and accuracy sklearn evaluation metrics
+
+    and fetches a summary report.
+    """
     from sklearn import metrics
 
     report = metrics.classification_report(test_y, preds)
     precision = metrics.precision_score(test_y, preds, average='weighted')
     f1 = metrics.f1_score(test_y, preds, average='weighted')
     accuracy = metrics.accuracy_score(test_y, preds)
-
-    if log:
-        logger.info("Intent Evaluation Results")
-        logger.info("F1-Score:  {}".format(f1))
-        logger.info("Precision: {}".format(precision))
-        logger.info("Accuracy:  {}".format(accuracy))
-        logger.info("Classification report: \n{}".format(report))
 
     return report, precision, f1, accuracy
 
@@ -145,7 +153,7 @@ def evaluate_intents(targets, predictions):  # pragma: no cover
     targets, predictions = remove_empty_intent_examples(targets, predictions)
     logger.info("Intent Evaluation: Only considering those {} examples that "
                 "have a defined intent out of {} examples".format(targets.size, num_examples))
-    get_evaluation_table(targets, predictions)
+    log_evaluation_table(targets, predictions)
 
     cnf_matrix = confusion_matrix(targets, predictions)
     plot_confusion_matrix(cnf_matrix, classes=unique_labels(targets, predictions),
@@ -158,6 +166,7 @@ def merge_labels(aligned_predictions, extractor=None):
     """Concatenates all labels of the aligned predictions.
 
     Takes the aligned prediction labels which are grouped for each message
+
     and concatenates them.
     """
     if extractor:
@@ -190,7 +199,7 @@ def evaluate_entities(targets, predictions, tokens, extractors):  # pragma: no c
         merged_predictions = merge_labels(aligned_predictions, extractor)
         merged_predictions = substitute_labels(merged_predictions, "O", "no_entity")
         logger.info("Evaluation for entity extractor: {} ".format(extractor))
-        get_evaluation_table(merged_targets, merged_predictions)
+        log_evaluation_table(merged_targets, merged_predictions)
 
 
 def is_token_within_entity(token, entity):