+++ rasa_nlu/server.py
@@ -219,13 +219,17 @@ class RasaNLU(object):
     @check_cors
     def evaluate(self, request):
         data_string = request.content.read().decode('utf-8', 'strict')
-        kwargs = {key.decode('utf-8', 'strict'): value[0].decode('utf-8', 'strict')
-                  for key, value in request.args.items()}
+        params = {
+            key.decode('utf-8', 'strict'): value[0].decode('utf-8', 'strict')
+            for key, value in request.args.items()
+        }
         request.setHeader('Content-Type', 'application/json')
 
         try:
             request.setResponseCode(200)
-            response = self.data_router.evaluate(data_string, kwargs)
+            response = self.data_router.evaluate(data_string,
+                                                 params.get('project'),
+                                                 params.get('model'))
             return simplejson.dumps(response)
         except Exception as e:
             request.setResponseCode(500)